+++ docs/development/changelog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - RX gate
+- CRX gate
+- CRZ gate
 
 ## [0.1.6] - 2026-01-09